Refactor Potter app to bloc architecture
