+++ lib/features/characters/widgets/info_row.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class CharacterInfoRow extends StatelessWidget {
+  const CharacterInfoRow(this.title, this.value, {super.key});
+
+  final String title;
+  final String value;
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 6),
+      child: Row(
+        children: [
+          SizedBox(
+            width: 100,
+            child: Text(
+              title,
+              style: textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.w600),
+            ),
+          ),
+          const SizedBox(width: 12),
+          Expanded(child: Text(value)),
+        ],
+      ),
+    );
+  }
+}
